While Loop Practice
Figured out how to print out a list depending on what two integers the user inputs into the command prompt.

While loop checks the inputs the user entered and makes sure its 1 - 100 and nothing above or negative. If the numbers are equaled to each other it will ask the user to input two new numbers again.

if / else statement checks if the user's inputs are greater than or less than in order to output the list of numbers in order.